added svatky, fixed kurzy, fixed damejidlo

+++ app/src/main/java/io/antelli/plugin/seznam/pocasi/entity/Daily.kt
@@ -12,7 +12,7 @@ data class Daily(
         @SerializedName("localDate")
     val localDate: Date,
         @SerializedName("snowPrecip")
-    val snowPrecip: Int,
+    val snowPrecip: Double,
         @SerializedName("sunrise")
     val sunrise: String,
         @SerializedName("sunset")
